requirements.txt
alembic==1.5.2
amqp==5.0.2
appdirs==1.4.4
atlastk==0.12.4
bcrypt==3.2.0
billiard==3.6.3.0
celery==5.0.4
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
cryptography==3.3.1
ecdsa==0.14.1
fastapi==0.63.0
fastapi-contrib==0.2.9
gunicorn==20.0.4
h11==0.12.0
httptools==0.1.1
idna==2.10
jaeger-client==4.4.0
Jinja2==2.11.2
kombu==5.0.2
Mako==1.1.4
MarkupSafe==1.1.1
mongoengine==0.22.1
motor==2.3.0
opentracing==2.4.0
passlib==1.7.4
prompt-toolkit==3.0.9
psycopg2==2.8.6
pyasn1==0.4.8
pycparser==2.20
pycrypto==2.6.1
pydantic==1.7.3
pyee==8.1.0
pymongo==3.11.2
pyppeteer==0.2.5
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
python-jose==3.2.0
python-multipart==0.0.5
pytz==2020.5
PyYAML==5.4.1
redis==3.5.3
requests==2.25.1
rsa==4.7
simple-crypt==4.1.7
six==1.15.0
SQLAlchemy==1.3.22
starlette==0.13.6
threadloop==1.0.2
thrift==0.13.0
tornado==6.1
tqdm==4.56.0
ujson==1.35
urllib3==1.26.2
uvicorn==0.13.3
uvloop==0.14.0
vine==5.0.0
watchgod==0.6
wcwidth==0.2.5
websockets==8.1
